+++ packages/blocky-common/src/index.ts
@@ -4,3 +4,8 @@ export * from "./disposable";
 export * from "./events";
 export * from "./lazy";
 export * from "./stack";
+export * from "./position";
+export * from "./text";
+export * from "./character";
+export * from "./object";
+export * from "./hash";
