+++ packages/blocky-core/src/block/basic.ts
@@ -1,5 +1,5 @@
-import { type IDisposable } from "blocky-common/es/disposable";
-import { type Position } from "blocky-common/es/position";
+import { type IDisposable, type Position } from "blocky-common/es";
+import { elem } from "blocky-common/es/dom";
 import { type HTMLConverter } from "@pkg/helper/htmlConverter";
 import {
   type BlockyNode,
@@ -11,7 +11,6 @@ import {
 } from "blocky-data";
 import { type Editor } from "@pkg/view/editor";
 import { type EditorController } from "@pkg/view/controller";
-import { elem } from "blocky-common/es/dom";
 
 export interface BlockDidMountEvent {
   element: HTMLElement;
